requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlalchemy
alembic
psycopg2-binary
pytest
httpx
python-dotenv
pytest-asyncio
pytest-cov
black
isort
mypy
flake8
python-dateutil
